+++ ui/components/providers/radio-group-provider.tsx
@@ -12,6 +12,7 @@ import {
   AlibabaCloudProviderBadge,
   AWSProviderBadge,
   AzureProviderBadge,
+  CloudflareProviderBadge,
   GCPProviderBadge,
   GitHubProviderBadge,
   IacProviderBadge,
@@ -73,6 +74,11 @@ const PROVIDERS = [
     value: "alibabacloud",
     label: "Alibaba Cloud",
     badge: AlibabaCloudProviderBadge,
+  },
+  {
+    value: "cloudflare",
+    label: "Cloudflare",
+    badge: CloudflareProviderBadge,
   },
   {
     value: "openstack",